Decoder modifications in preparation for merging more of the encoder and  decoder code. As a side effect, this speeds things up 3.2% on x86-32 and 11.1% on x86-64.
svn path=/branches/theora-thusnelda/; revision=15977

+++ lib/dec/decint.h
@@ -58,15 +58,15 @@ struct th_dec_ctx{
   oc_huff_node        *huff_tables[TH_NHUFFMAN_TABLES];
   /*The index of one past the last token in each plane for each coefficient.
     The final entries are the total number of tokens for each coefficient.*/
-  int                  ti0[3][64];
+  ptrdiff_t            ti0[3][64];
   /*The index of one past the last extra bits entry in each plane for each
      coefficient.
     The final entries are the total number of extra bits entries for each
      coefficient.*/
-  int                  ebi0[3][64];
+  ptrdiff_t            ebi0[3][64];
   /*The number of outstanding EOB runs at the start of each coefficient in each
      plane.*/
-  int                  eob_runs[3][64];
+  ptrdiff_t            eob_runs[3][64];
   /*The DCT token lists.*/
   unsigned char      **dct_tokens;
   /*The extra bits associated with DCT tokens.*/
